==> Capstone_3/Notebooks/2_Data-Wrangling_Capstone-3.ipynb <==
### Data Wrangling

In this notebook, will go through the data wrangling process for our project. The sponser, WebFilms, has provided two sources of data for us to use:

* What they currently have in their database for movie ratings (Recommendation System.csv)
* The Movie ID and Title keys (Movie_Id_Titles).

The sponser is clear that they don't want us using any other data than the data provided so we will work with the data that is provided.

# First, let's import relevant packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

# Second, let's first take a look at the data provided regarding ratings
#import the csv into a Dataframe
ratings_df = pd.read_csv('Recommendation System.csv')

#Let's look at the DataFrame
ratings_df.head()

According to the documentation, we have the following categories of information:
* user_id: The unique identifier for a user
* item_id: The unique identifier for a movie
* rating: the rating that a specific user gave a specific movie on a scale of 1 to 5
* timestamp: when in time the rating was given

This means we'll need to apply column labels to our dataframe

#First, let's create a list of our column labels
column_names = ['user_id', 'movie_id', 'rating', 'timestamp']

#Second, let's apply our column names to the dataframe
ratings_df.columns = column_names

#Check to see that everything worked
ratings_df.head()

The first thing we should consider is whether there are any empty elements in our table or anything that may make it difficult to conduct exploratory data analysis (EDA). Therefore, let's explore our data a little to get a sense of what we have.

#Let's get some summary information about our dataframe
ratings_df.describe()

ratings_df.info()

There doesn't appear to be any evidence that we are missing information or any indication that we should be concerned at a cursory level about the data. That said, let's calculate the number of unique entries for user_id and movie_id to get a sense of how many unique elements are in each column

#Count number of unique users
print(f'Number of unique users: {len(ratings_df.user_id.unique())}')

#Count number of unique movies
print(f'Number of unique movies: {len(ratings_df.movie_id.unique())}')

This would imply that we have 944 Users and 1682 Movies in our database, which lines up with the information provided by the sponsor. Therefore we should be confident that the data is accurate for these two columns. The next step is to check the ratings for the rows to make sure they are integers between 1-5. Although the descriptive statistics would imply they are, can't hurt to be a little extra careful.

#Get all the unique elements in the rating column
print(ratings_df.rating.unique())

This gives us confidence that our data for ratings is proper. One last bit of information is that the sponsor provided the movie title associated with the movie_ids in a different table. While it's not clear this will be helpful, we can add them to the dataframe for now and always drop it if necesssary.

#Load table with movie_id and title
movie_table = pd.read_table('Movie_Id_Titles', delimiter=',', header = 0)

#Let's check to make sure everything came in properly
movie_table.head()

#Let's change the column headers to match up with our already established data frame and to make them slightly more expressive
column_names_2 = ['movie_id', 'movie_title']

movie_table.columns = column_names_2

#Check to make sure everything worked properly
movie_table.head()

#Now we can merge our dataframes together into a single dataframe with completely information
ratings_df_complete = ratings_df.merge(movie_table, how='left', on='movie_id')

#Check to make sure everything worked
ratings_df_complete.head(15)

#Let's check that if a movie_id is the same, you get the same title
ratings_check = ratings_df_complete.loc[ratings_df_complete.movie_id == 302]
ratings_check.head()

Alright! Based on table and our checks, it looks we should be good to go on to EDA The last step would be to export our table so we can use it later

#Export as a CSV so we can begin using it
ratings_df_complete.to_csv('Ratings_Table_COMPLETE.csv')

That should complete this portion of the project! Next, we'll move onto EDA which may need to be modified a little given that we are working on a recommendation system!

Note on outliers: For a data set like this, where outliers need to be evaluted on either a row-based metric (i.e., did the user rate a bunch of movies but a specific movie rated by that users is substantially different than the other movies the user rated) or a column-based metric (i.e., a movie has gotten an in ordinate amount of ratings and a few of the ratings fall outside the range of the many ratings), it's going to be tough to really parse out whether an "outlier" is of value or something else. This is because the features we are evaluating are going to be explored in the context of "collaborative filting" where each user has preferences and it is perfectly valid for them to have "outlier" observations. During the rest of the process, we may come across items that need to be considered, but for now, it's tough to deduce if anything is genuinely an outlier.
